Fixed a few defects in the collator and expression which were using longs instead of floats

+++ metrics/model/src/main/java/org/scriptbox/metrics/query/exp/TotalPerSecondQueryExp.java
@@ -26,7 +26,7 @@ public class TotalPerSecondQueryExp implements MetricQueryExp {
 		return collator.collate(false, new ParameterizedRunnableWithResult<Metric, MetricRange>() {
 			public Metric run(MetricRange range) {
 				Iterator<Metric> iter = range.getIterator(ctx.getChunk());
-				long total = 0;
+				float total = 0;
 				while (iter.hasNext()) {
 					Metric metric = iter.next();
 					total += metric.getValue();
