+++ metrics/model/src/main/java/org/scriptbox/metrics/query/exp/TotalQueryExp.java
@@ -29,7 +29,7 @@ public class TotalQueryExp implements MetricQueryExp {
 		return collator.collate(false, new ParameterizedRunnableWithResult<Metric, MetricRange>() {
 			public Metric run(MetricRange range) {
 				Iterator<Metric> iter = range.getIterator(ctx.getChunk());
-				long total = 0;
+				float total = 0;
 				while (iter.hasNext()) {
 					Metric metric = iter.next();
 					total += metric.getValue();
